+++ apps/teams/urls.py
@@ -9,8 +9,6 @@ urlpatterns = [
     # invitation acceptance views
     path("invitation/<slug:invitation_id>/", views.accept_invitation, name="accept_invitation"),
     path("invitation/<slug:invitation_id>/signup/", views.SignupAfterInvite.as_view(), name="signup_after_invite"),
-    path("sudo/<slug:team_slug>/", views.acquire_superuser_powers, name="sudo"),
-    path("sudo/<slug:team_slug>/release/", views.release_superuser_powers, name="release_sudo"),
 ]
 
 team_urlpatterns = (
