+++ apps/teams/views/__init__.py
@@ -1,4 +1,3 @@
 from .invitation_views import *  # noqa F401
 from .manage_team_views import *  # noqa F401
 from .membership_views import *  # noqa F401
-from .superuser import *  # noqa F401
