+++ apps/web/superuser_utils.py
@@ -5,33 +5,33 @@ from django.utils import timezone
 EXPIRY = 300
 
 
-def apply_temporary_superuser_access(request, team_slug):
+def apply_temporary_superuser_access(request, slug):
     remove_expired_temporary_superuser_access(request)
-    if has_temporary_superuser_access(request, team_slug):
+    if has_temporary_superuser_access(request, slug):
         return
 
     elevated_privileges = request.session.get("elevated_privileges", [])
     expire = timezone.now() + timedelta(seconds=EXPIRY)
-    elevated_privileges.append((team_slug, int(expire.timestamp())))
+    elevated_privileges.append((slug, int(expire.timestamp())))
     request.session["elevated_privileges"] = elevated_privileges
 
 
-def has_temporary_superuser_access(request, team_slug):
+def has_temporary_superuser_access(request, slug):
     elevated_privileges = request.session.get("elevated_privileges", [])
     now = int(timezone.now().timestamp())
-    has_access = any(team == team_slug and expire > now for team, expire in elevated_privileges)
+    has_access = any(team == slug and expire > now for team, expire in elevated_privileges)
     if not has_access:
-        remove_temporary_superuser_access(request, team_slug)
+        remove_temporary_superuser_access(request, slug)
     return has_access
 
 
-def remove_temporary_superuser_access(request, team_slug):
+def remove_temporary_superuser_access(request, slug):
     """This removes access to the specific team and also expired access."""
 
     elevated_privileges = request.session.get("elevated_privileges", [])
     now = int(timezone.now().timestamp())
     request.session["elevated_privileges"] = [
-        (team, expire) for team, expire in elevated_privileges if team != team_slug or expire <= now
+        (team, expire) for team, expire in elevated_privileges if team != slug or expire <= now
     ]
