+++ templates/web/components/notices.html
@@ -2,7 +2,7 @@
   <div class="alert bg-red-300 dark:bg-red-800 shadow-md !rounded-none text-warning-content">
     You are operating with escalated privileges. Release access:
     {% for slug in temporary_superuser_access %}
-      <a href="{% url 'teams:release_sudo' slug %}" class="btn btn-warning btn-xs">
+      <a href="{% url 'web:release_sudo' slug %}" class="btn btn-warning btn-xs">
         {{ slug }}
       </a>
     {% endfor %}
